Add option to use openldap as an identity backend
Keystone supports using ldap as an identity backend for storing users and groups.
This commit makes changes to install openldap and modify the keystone and haproxy
configuration to support this change.
Openldap is installed in master master replication on two nodes using haproxy for load
balancing. LDAP driver does not support multiple domains, due to this, keystone is
configured to use domain specific drivers. Users in default domain are stored in LDAP,
while users belonging to other domains are stored in sql backend. Heat needs this feature
as the heat stack users are created in a separate domain called the heat-domain.

Main reason for the change is to support compliance requirements for user passwords by
the way of using password policies in openldap that support following things.
  - Syntax check for password strength.
  - Account locking on multiple tries with wrong password.
  - Password encryption.
  - Expire a password after a certain duration.
  - Disallow password change before a certain duration.

Certain users like service users(nova, glance, cinder etc) can be selectively exempted
by applying a password policy that has no rules.